+++ src/MystiqueValue.php
@@ -3,12 +3,7 @@
 namespace Altivebir\Mystique;
 
 use ProcessWire\Language;
-use ProcessWire\Mystique;
-use ProcessWire\Field;
-use ProcessWire\Page;
 use ProcessWire\WireData;
-use ProcessWire\WireException;
-use ProcessWire\InputfieldMystique;
 
 /**
  * Class MystiqueValue
@@ -16,82 +11,28 @@ use ProcessWire\InputfieldMystique;
  * @author			: İskender TOTOĞLU, @ukyo (community), @trk (Github)
  * @website			: https://www.altivebir.com
  *
- * @property $__json
- * @property $__name
- * @property $__path
- * @property $__resource
- *
  * @package Altivebir\Mystique
  */
 class MystiqueValue extends WireData
 {
     /**
-     * @var Mystique $Mystique
+     * @var array
      */
-    protected $Mystique;
-
-    /**
-     * @var MystiqueFormManager
-     */
-    private $manager;
-
-    /**
-     * @var Page
-     */
-    private $page;
-
-    /**
-     * @var InputfieldMystique
-     */
-    private $field;
+    protected $languageFields;
 
     /**
      * @inheritDoc
-     *
-     * @param Page $page
-     * @param InputfieldMystique $field
-     * @throws WireException
      */
-    public function __construct(Page $page, Field $field)
+    public function __construct(array $values = [], array $languageFields = [])
     {
         parent::__construct();
 
-        $this->Mystique = $this->wire("modules")->get("Mystique");
-        $this->page = $page;
-        $this->field = $field;
+        $this->languageFields = $languageFields;
 
-        if($this->field->useJson && $this->field->jsonString || $field->resource) {
+        foreach ($values as $name => $value) {
+            
+            $this->set($name, $value);
 
-            $this->manager = new MystiqueFormManager($field, $page);
-
-            if($this->field->useJson && $this->field->jsonString) {
-                $resource = json_decode($field->jsonString, true);
-            } else {
-                // $resource =  $this->Mystique->resource($this->field->resource);
-                $resource =  $this->Mystique->getResource($this->field->resource);
-            }
-
-            // Set default values
-            foreach ($this->manager->inputFields as $name => $value) {
-                if(in_array($name, $this->manager->languageFields)) {
-                    $this->set($name, $value);
-                    foreach ($this->languages ?: [] as $language) {
-                        if ($language->isDefault()) {
-                            continue;
-                        } else {
-                            $this->set($name . $language->id, $value);
-                        }
-                    }
-                } else {
-                    $this->set($name, $value);
-                }
-            }
-
-            $this->set('__json', json_encode($resource));
-            $this->set('__name', isset($resource['__name']) ? $resource['__name'] : '');
-            $this->set('__path', isset($resource['__path']) ? $resource['__path'] : '');
-        } else {
-            throw new WireException("You need to select a resource and save field before start to use Mystique.");
         }
     }
 
@@ -100,7 +41,7 @@ class MystiqueValue extends WireData
      */
     public function get($key)
     {
-        if(in_array($key, $this->manager->languageFields)) {
+        if (in_array($key, $this->languageFields)) {
             $user = $this->user;
             $language = $user ? $user->language : null;
             if($language instanceof Language && !$language->isDefault) {
@@ -117,53 +58,5 @@ class MystiqueValue extends WireData
     public function set($key, $value)
     {
         return parent::set($key, $value);
-    }
-
-    /**
-     * @return Page
-     */
-    public function getPage()
-    {
-        return $this->page;
-    }
-
-    public function getResource()
-    {
-        return $this->get('__resource');
-    }
-
-    public function getPath()
-    {
-        return $this->get('__path');
-    }
-
-    /**
-     * Return data as array
-     *
-     * @return array
-     */
-    public function array(): array
-    {
-        return $this->data;
-    }
-
-    /**
-     * Return data as string
-     *
-     * @return string
-     */
-    public function json(): string
-    {
-        return json_encode($this->array());
-    }
-
-    /**
-     * Return json data as string
-     *
-     * @return string
-     */
-    public function __toString(): string
-    {
-        return $this->json();
     }
 }